+++ openmetadata-spec/src/main/resources/json/schema/api/createEventPublisherJob.json
@@ -1,0 +1,41 @@
+{
+  "$id": "https://open-metadata.org/schema/entity/feed/createEventPublisherJob.json",
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "title": "CreateEventPublisherJob",
+  "description": "This schema defines Event Publisher Run Result.",
+  "type": "object",
+  "javaType": "org.openmetadata.schema.api.CreateEventPublisherJob",
+  "properties": {
+    "name": {
+      "description": "Name of the Job",
+      "type": "string"
+    },
+    "publisherType": {
+      "$ref": "../settings/eventPublisherJob.json#/definitions/publisherType"
+    },
+    "runMode": {
+      "$ref": "../settings/eventPublisherJob.json#/definitions/runMode"
+    },
+    "entities": {
+      "description": "List of Entities to Reindex",
+      "type": "array",
+      "items": {
+        "type": "string"
+      },
+      "default": ["all"],
+      "uniqueItems": true
+    },
+    "recreateIndex": {
+      "description": "This schema publisher run modes.",
+      "type": "boolean",
+      "default": false
+    },
+    "batchSize": {
+      "description": "Maximum number of events sent in a batch (Default 10).",
+      "type": "integer",
+      "default": 100
+    }
+  },
+  "required": ["name", "publisherType","runMode"],
+  "additionalProperties": false
+}
